requirements.txt
# Core
Django==5.0.6
django-environ==0.11.2
gunicorn==22.0.0
gevent==24.2.1
django_tenants==3.6.1
celery==5.4.0
cryptography==42.0.8
python-dotenv==1.0.1

# For the persistence stores
psycopg2-binary==2.9.9

# Rest apis
djangorestframework==3.15.2
djangorestframework_simplejwt==5.3.1
django-cors-headers==4.4.0
drf-yasg==1.21.7
whitenoise==6.7.0
django-filter==24.2
